pyproject.toml
[tool.poetry]
name = "AntonIA"
version = "0.1.1"
description = "This is a pipeline setup to generate good morning images"
authors = ["Carmoldu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^2.1.0"
dotenv = "^0.9.9"
pandas = "^2.3.3"
pillow = "^11.3.0"
pyarrow = "^21.0.0"
pyyaml = "^6.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ipykernel = "^6.30.0"
pytest-cov = "^6.2.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"


[tool.pytest.ini_options]
pythonpath = "src"
